+++ ui/upgrade_menu.gd
@@ -16,6 +16,9 @@ var upgrade_pool: Array[UpgradeResource] = [
 	preload("res://data/upgrades/attack_speed.tres"),
 	preload("res://data/upgrades/heal.tres"),
 	preload("res://data/upgrades/max_health.tres"),
+	preload("res://data/upgrades/unlock_sniper.tres"),
+	preload("res://data/upgrades/unlock_orbiter.tres"),
+	preload("res://data/upgrades/unlock_healer.tres"),
 ]
 
 # --- SIGNALS ---